LCAM-1191: Addressed review comments
